Update classes for better interaction with database
This commit introduces several structural updates to classes to improve their interaction with the database. Lombok library was added to several classes, replacing getter and setter methods for better readability and efficiency.

The Borrow class was updated to include a borrowId identifier to keep track of activity. This change required updates to related classes and methods. Modifications in User and Book classes were introduced for enhanced data retrieval and modification.

In the Library class, Lombok annotations were again used to replace getters and setters. A major change was refactoring initialization data load with queries on the database.

Updates in the Requetes class further optimize data retrieval and set up from the database, with exception management to alert on errors.

Furthermore, updates in the data.sql file changed assigned roles in the user table.

All these changes make the project more aligned with proper practices for interaction with a relational database. It improves code readability, and sets a stronger foundation for future database integration.